Upgrade smithy to 1.53.0
