+++ codegen/core/src/test/java/software/amazon/smithy/java/codegen/utils/TestJavaCodegenPlugin.java
@@ -9,9 +9,9 @@ import software.amazon.smithy.build.PluginContext;
 import software.amazon.smithy.build.SmithyBuildPlugin;
 import software.amazon.smithy.codegen.core.directed.CodegenDirector;
 import software.amazon.smithy.java.codegen.CodeGenerationContext;
+import software.amazon.smithy.java.codegen.DefaultTransforms;
 import software.amazon.smithy.java.codegen.JavaCodegenIntegration;
 import software.amazon.smithy.java.codegen.JavaCodegenSettings;
-import software.amazon.smithy.java.codegen.transforms.DefaultTransforms;
 import software.amazon.smithy.java.codegen.writer.JavaWriter;
 
 public class TestJavaCodegenPlugin implements SmithyBuildPlugin {
@@ -29,13 +29,9 @@ public class TestJavaCodegenPlugin implements SmithyBuildPlugin {
         runner.directedCodegen(new TestJavaCodegen());
         runner.fileManifest(context.getFileManifest());
         runner.service(settings.service());
-
-        var model = DefaultTransforms.transform(context.getModel(), settings);
-
-        runner.model(model);
+        runner.model(context.getModel());
         runner.integrationClass(JavaCodegenIntegration.class);
-        runner.performDefaultCodegenTransforms();
-        runner.createDedicatedInputsAndOutputs();
+        DefaultTransforms.apply(runner, settings);
         runner.run();
     }
 }
